requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-cors-headers==2.4.0
django-tastypie==0.14.2
djangorestframework==3.9.0
gunicorn==19.6.0
idna==2.8
pbr==5.1.1
pipenv==2018.11.26
pipenv-to-requirements==0.6.1
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
virtualenv==16.1.0
virtualenv-clone==0.4.0
